+++ client/src/components/carousel/CarouselLogic.tsx
