+++ client/src/components/concert-preview/ConcertPreview.tsx
@@ -1,5 +1,5 @@
 import { styled } from 'styled-components';
-import { ButtonWithArrowDark } from '../Buttons/Buttons';
+import { ButtonWithArrowDark } from '../buttons/Buttons';
 import Img from '../.././images/우리사랑이대로.jpeg';
 
 export default function Concertpreview() {
