+++ client/src/components/slogan/Slogan.tsx
